Widen the read seam to u64, so a map may be larger than a u32
`obc_formats::io::ByteSource` is the whole tree's read interface — every format
parser reaches its bytes through it and nothing else — so its width was the
largest file anything here could open. It was `read_at(offset: u32)` /
`len() -> u32`, which put the practical wall at 4 GiB no matter that OBCM v14's
interior is 64 GiB. `obcm_assemble::FILE_CEILING` was the `min` of the two, and
the readable one bound.

The seam and everything behind it now count file offsets in 64 bits: both
implementors and both caches in `obc-reader`, the flat store's `StoreSource`
(whose saturation-at-`u32::MAX` dies — the store's lengths were always u64), the
FAT arm, the sim's and the assembler's sources, and the web bridge's slot reads,
which serve a sunk shard out of host storage rather than out of the address
space. `FILE_CEILING`'s readable arm follows, so the format wall binds at 64 GiB
and `CORE_WARN` follows it as the proportion it was written to be.

Three things stay 32-bit and say why where they are: `ByteSink::patch_at` (OBCR
route files, whose own offsets are `uint32`), each format's own internal offsets
(OBCR, OBCT, OBCW), and every buffer *length*, which is bounded by the host's
address space rather than by a format — a different fact that happens to share a
number on wasm32.

OBCA §5.2's `Bytes` is `uint32` and is deliberately not widened, so a set's
member keeps the old wall: `SET_SHARDS_CEILING` names it at plan time, where
refusing costs nothing, rather than letting the split path write a 6 GiB shard
and fail at manifest emit. The single file §5.5 plans first carries no manifest
record and gets the whole ceiling.

`tests/far_offsets.rs` is the capability: a real two-LOD map relocated so every
section begins past 5 GiB, parsed and decoded feature-for-feature against the
same map at low offsets. It could not have been written before this change.

Resident cost on the board is +1,224 B, 69 % of it `SET_SHARDS` — a table
FS7.5b2 deletes. `IndexBlock` keeps a `u32` **block number** rather than a `u64`
byte offset, and that is measured rather than preferred: the `u64` took
`MapCacheInner` from 4- to 8-aligned, LLVM stopped folding its `zeroed()` into a
`.bss` memset at the board's placement site, and the boot task's frame went
7,456 -> 74,240 B against a 49 KB residual stack. Both board profiles' baselines
are re-approved with the itemization.

+++ apps/obc-web-assemble/src/lib.rs
@@ -246,8 +246,10 @@ mod web {
             JsSink::taken(self.write.call2(&JsValue::NULL, &JsValue::from_f64(slot as f64), &src), "write")
         }
 
-        fn read_at(&self, slot: usize, offset: u32, into: &mut [u8]) -> Result<(), String> {
+        fn read_at(&self, slot: usize, offset: u64, into: &mut [u8]) -> Result<(), String> {
             // SAFETY: as in `JsReads::read` — a per-call view, filled and dropped inside the call.
+            // A sunk shard's offset can pass 4 GiB since the read seam widened; `f64` carries it
+            // exactly to 2^53, far past the 64 GiB §1.1 lets a file reach.
             let dest = unsafe { js_sys::Uint8Array::view_mut_raw(into.as_mut_ptr(), into.len()) };
             JsSink::taken(
                 self.read_at.call3(
@@ -373,7 +375,7 @@ mod web {
     }
 
     impl CellReads for JsReads {
-        fn read(&self, slot: usize, offset: u32, buf: &mut [u8]) -> Result<(), String> {
+        fn read(&self, slot: usize, offset: u64, buf: &mut [u8]) -> Result<(), String> {
             // SAFETY: `view_mut_raw` aliases linear memory and is invalidated by anything that grows
             // it. This one is made, passed, and dropped inside a single synchronous JS call that
             // does nothing but fill it — no Rust allocation can run in between, and the callback is
